fix: fixed email verification
